store data matrix X as a contiguous float32 array (#143)

+++ scvi/metrics/imputation.py
@@ -29,7 +29,9 @@ def imputation(infer, name, rate=0.1, n_samples=1, n_epochs=1, corruption="unifo
             range(len(i)), int(np.floor(rate * len(i))), replace=False
         )
         i, j = i[ix], j[ix]
-        corrupted_data[i, j] = np.random.binomial(n=corrupted_data[i, j], p=0.2)
+        corrupted_data[i, j] = np.random.binomial(
+            n=corrupted_data[i, j].astype(np.int64), p=0.2
+        )
 
     infer.gene_dataset = gene_dataset = GeneExpressionDataset(
         *GeneExpressionDataset.get_attributes_from_matrix(
